mongo embedded verticle

+++ ext-mongo/src/main/resources/ext-mongo-js/mongo_service.js
@@ -30,112 +30,150 @@ var MongoService = function(j_val) {
   var that = this;
 
   this.save = function(collection, document, writeConcern, resultHandler) {
-    j_mongoService.save(collection, utils.convJSObjectToJsonObject(document), writeConcern, function(ar) {
+    var __args = arguments;
+    if (__args.length === 4 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'string' && typeof __args[3] === 'function') {
+      j_mongoService.save(collection, utils.convJSObjectToJsonObject(document), writeConcern, function(ar) {
       if (ar.succeeded()) {
         resultHandler(ar.result(), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.insert = function(collection, document, writeConcern, resultHandler) {
-    j_mongoService.insert(collection, utils.convJSObjectToJsonObject(document), writeConcern, function(ar) {
+    var __args = arguments;
+    if (__args.length === 4 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'string' && typeof __args[3] === 'function') {
+      j_mongoService.insert(collection, utils.convJSObjectToJsonObject(document), writeConcern, function(ar) {
       if (ar.succeeded()) {
         resultHandler(ar.result(), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.update = function(collection, query, update, writeConcern, upsert, multi, resultHandler) {
-    j_mongoService.update(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(update), writeConcern, upsert, multi, function(ar) {
+    var __args = arguments;
+    if (__args.length === 7 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'object' && typeof __args[3] === 'string' && typeof __args[4] ==='boolean' && typeof __args[5] ==='boolean' && typeof __args[6] === 'function') {
+      j_mongoService.update(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(update), writeConcern, upsert, multi, function(ar) {
       if (ar.succeeded()) {
         resultHandler(null, null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.find = function(collection, query, fields, sort, limit, skip, resultHandler) {
-    j_mongoService.find(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(fields), utils.convJSObjectToJsonObject(sort), limit, skip, function(ar) {
+    var __args = arguments;
+    if (__args.length === 7 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'object' && typeof __args[3] === 'object' && typeof __args[4] ==='number' && typeof __args[5] ==='number' && typeof __args[6] === 'function') {
+      j_mongoService.find(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(fields), utils.convJSObjectToJsonObject(sort), limit, skip, function(ar) {
       if (ar.succeeded()) {
         resultHandler(utils.convListSetJson(ar.result()), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.findOne = function(collection, query, fields, resultHandler) {
-    j_mongoService.findOne(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(fields), function(ar) {
+    var __args = arguments;
+    if (__args.length === 4 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'object' && typeof __args[3] === 'function') {
+      j_mongoService.findOne(collection, utils.convJSObjectToJsonObject(query), utils.convJSObjectToJsonObject(fields), function(ar) {
       if (ar.succeeded()) {
         resultHandler(utils.convJsonToJS(ar.result()), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.delete = function(collection, query, writeConcern, resultHandler) {
-    j_mongoService.delete(collection, utils.convJSObjectToJsonObject(query), writeConcern, function(ar) {
+    var __args = arguments;
+    if (__args.length === 4 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'string' && typeof __args[3] === 'function') {
+      j_mongoService.delete(collection, utils.convJSObjectToJsonObject(query), writeConcern, function(ar) {
       if (ar.succeeded()) {
         resultHandler(null, null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.createCollection = function(collectionName, resultHandler) {
-    j_mongoService.createCollection(collectionName, function(ar) {
+    var __args = arguments;
+    if (__args.length === 2 && typeof __args[0] === 'string' && typeof __args[1] === 'function') {
+      j_mongoService.createCollection(collectionName, function(ar) {
       if (ar.succeeded()) {
         resultHandler(null, null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.getCollections = function(resultHandler) {
-    j_mongoService.getCollections(function(ar) {
+    var __args = arguments;
+    if (__args.length === 1 && typeof __args[0] === 'function') {
+      j_mongoService.getCollections(function(ar) {
       if (ar.succeeded()) {
         resultHandler(ar.result(), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.dropCollection = function(collection, resultHandler) {
-    j_mongoService.dropCollection(collection, function(ar) {
+    var __args = arguments;
+    if (__args.length === 2 && typeof __args[0] === 'string' && typeof __args[1] === 'function') {
+      j_mongoService.dropCollection(collection, function(ar) {
       if (ar.succeeded()) {
         resultHandler(null, null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.runCommand = function(collection, command, resultHandler) {
-    j_mongoService.runCommand(collection, utils.convJSObjectToJsonObject(command), function(ar) {
+    var __args = arguments;
+    if (__args.length === 3 && typeof __args[0] === 'string' && typeof __args[1] === 'object' && typeof __args[2] === 'function') {
+      j_mongoService.runCommand(collection, utils.convJSObjectToJsonObject(command), function(ar) {
       if (ar.succeeded()) {
         resultHandler(utils.convJsonToJS(ar.result()), null);
       } else {
         resultHandler(null, ar.cause());
       }
     });
+    } else utils.invalidArgs();
   };
 
   this.start = function() {
-    j_mongoService.start();
+    var __args = arguments;
+    if (__args.length === 0) {
+      j_mongoService.start();
+    } else utils.invalidArgs();
   };
 
   this.stop = function() {
-    j_mongoService.stop();
+    var __args = arguments;
+    if (__args.length === 0) {
+      j_mongoService.stop();
+    } else utils.invalidArgs();
   };
+
+  this._vertxgen = true;
 
   // Get a reference to the underlying Java delegate
   this._jdel = function() {
@@ -145,11 +183,17 @@ var MongoService = function(j_val) {
 };
 
 MongoService.create = function(vertx, config) {
-  return new MongoService(JMongoService.create(vertx._jdel(), utils.convJSObjectToJsonObject(config)));
+  var __args = arguments;
+  if (__args.length === 2 && typeof __args[0] === 'object' && __args[0]._vertxgen && typeof __args[1] === 'object') {
+    return new MongoService(JMongoService.create(vertx._jdel(), utils.convJSObjectToJsonObject(config)));
+  } else utils.invalidArgs();
 };
 
 MongoService.createEventBusProxy = function(vertx, address) {
-  return new MongoService(JMongoService.createEventBusProxy(vertx._jdel(), address));
+  var __args = arguments;
+  if (__args.length === 2 && typeof __args[0] === 'object' && __args[0]._vertxgen && typeof __args[1] === 'string') {
+    return new MongoService(JMongoService.createEventBusProxy(vertx._jdel(), address));
+  } else utils.invalidArgs();
 };
 
 // We export the Constructor function
